pyproject.toml
[project]
name = "orange-hrm-playwright"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "allure-pytest>=2.15.3",
    "faker>=40.12.0",
    "playwright>=1.58.0",
    "pytest-base-url>=2.1.0",
    "pytest-playwright>=0.7.2",
]
[tool.pylint.messages_control]
disable = ["missing-module-docstring", "missing-function-docstring"]
[tool.pytest.ini_options]
addopts = [
  "--clean-alluredir",
  "--alluredir=allure-results",
  "--video=retain-on-failure",
  "--tracing=on"#retain-on-failure",
    ]

base_url = "https://opensource-demo.orangehrmlive.com/"
